Bailing on Bazel for now. Generate gqlgen app
I want to experiment with gqlgen and don't want to get bogged down right
now in trying to get Bazel to work with it. I was warned off trying from
a coworker that is familiar with Bazel. For now I want to learn gqlgen
so going to focus on that using the standard go tooling.

This commit deletes the Bazel stuff and generates a tutorial gqlgen
GraphQL application for todos.